(Follow-up to TAB-21) Consolidate version number validation and retrieval across types.
